backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
numpy>=1.26.0
pandas>=2.1.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
pydantic>=2.5.0
starlette>=0.27.0
python-docx>=0.8.11
pdfplumber>=0.10.3
Pillow>=10.1.0
pytesseract>=0.3.10 